requirements.txt
fastapi==0.104.1
packaging==23.1
httpx==0.25.2
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
sqlmodel==0.0.16
groq==0.8.0
openai>=1.30.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
